refactor: Clean code
